requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.13
django-braces==1.15.0
django-crispy-forms==2.2
django-dotenv==1.4.2
django-material==1.12.0
django-model-utils==4.5.1
django-phonenumber-field==8.0.0
django-seed==0.3.1
django-social-share==2.3.0
django-star-ratings==0.9.2
django-widget-tweaks==1.5.0
Faker==26.0.0
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
phonenumbers==8.13.40
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
standalone==1.0.1
swapper==1.4.0
toposort==1.10
twilio==9.2.4
typing_extensions==4.12.0
urllib3==2.2.2
yarl==1.9.8
